updated code

+++ main.py
@@ -1,7 +1,7 @@
 print("Calculate the area of wall.")
 feed = input("Enter the width of wall in metre: ")
-width = int(feed)
+width = float(feed)
 feed = input("Enter the height of the wall in metre: ")
-height = int(feed)
+height = float(feed)
 area = width * height
 print(f"The area of the wall is {area} square metres.")
